Melhorias e funções
